requirements.txt
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openpyxl==3.1.2
python-csv==0.0.13
subprocess.run==0.0.8
Pillow==9.5.0
tkintertable==1.3.3
numpy==1.23.5
face-recognition==1.3.0